Additions to config and scripting
