+++ mmg_toolbox/tkguis/misc/config.py
@@ -38,7 +38,7 @@ CONFIG_FILE = os.path.join(TMPDIR, TMPFILE)
 META_LIST = {
     # scan number and start_time included by default
     # name: format
-    'cmd': '{cmd|scan_command}'
+    'cmd': '{(cmd|scan_command)}'
 }
 
 REPLACE_NAMES = {
